Add a new command ledis and move all commands as the sub commands
